Weather included partially, hammering needed
Couldn't push using git. That's why archived the whole project and uploading here in this branch. There are some issues (weatheractivity.java and may be in the menu ) which needs to be fixed. I've done what I could for now. Download and check it out and let me know if it can be fixed.